Split common parts out of the server/client startup scripts
